Create container for pie chart
